demo-gl-antialiasing: refactor: Remove glFont package
The original version of this code had a built-in HUD that used Brad
Fish's 1998 helper library "glFont." Unfortunately that code is long
since forgotten by the world and internet, and support for it does not
exist.

glFont appears to have a Windows dependency at the binary-encoded font
level. Thus I am removing the HUD from this code altogether.

Signed-off-by: Rick Ramstetter <rick@anteaterllc.com>